+++ docs/components/theme/cache.tsx
@@ -11,16 +11,7 @@ const AppCache = (props: AppCacheProps) => {
 	// Props
 	const { dir, children } = props
 
-	return (
-		<AppRouterCacheProvider
-			options={{
-				enableCssLayer: true,
-				stylisPlugins: dir == 'rtl' ? [prefixer, rtlPlugin] : [prefixer]
-			}}
-		>
-			{children}
-		</AppRouterCacheProvider>
-	)
+	return <AppRouterCacheProvider options={{ stylisPlugins: dir == 'rtl' ? [prefixer, rtlPlugin] : [prefixer] }}>{children}</AppRouterCacheProvider>
 }
 
 export default AppCache